+++ playwright.config.ts
@@ -20,9 +20,5 @@ export default defineConfig({
   },
   projects: [
     { name: "chromium", use: { ...devices["Desktop Chrome"] } },
-    {
-      name: "mobile",
-      use: { ...devices["iPhone 13"], browserName: "chromium" },
-    },
   ],
 });